avm2: `pushshort` has an undocumented coercion step

+++ core/src/avm2/activation.rs
@@ -638,7 +638,7 @@ impl<'a, 'gc, 'gc_context> Activation<'a, 'gc, 'gc_context> {
     }
 
     fn op_push_short(&mut self, value: u32) -> Result<FrameControl<'gc>, Error> {
-        self.context.avm2.push(value);
+        self.context.avm2.push(value as i32);
         Ok(FrameControl::Continue)
     }
